tests/requirements.txt
pytest
ragger[speculos,ledgerwallet]>=1.11.4
pysha3>=1.0.0,<2.0.0
pynacl==1.3.0
